postgres_cdc: add integration tests
